pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flowbase"
version = "0.1.0"
description = "A platform for building tabular ML systems without infrastructure complexity"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Flowbase Team"}
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "duckdb>=0.9.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]
ml = [
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
]
cloud = [
    "boto3>=1.28.0",
    "trino>=0.328.0",
]
ui = [
    "streamlit>=1.28.0",
]

[project.scripts]
flowbase = "flowbase.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["flowbase*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
